--- src/Approvals/SFA.DAS.Approvals/Application/Apprentices/Queries/Apprenticeship/GetApprenticeshipsFilterValues/GetApprenticeshipsFilterValuesQuery.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-
-namespace SFA.DAS.Approvals.Application.Apprentices.Queries.Apprenticeship.GetApprenticeshipsFilterValues;
-
-public class GetApprenticeshipsFilterValuesQuery : IRequest<GetApprenticeshipsFilterValuesQueryResult>
-{
-    public long? EmployerAccountId { get; set; }
-    public long? ProviderId { get; set; }
-}
